# Code/gamesUP/docs/openapi.yaml
openapi: 3.0.1
info:
  title: GamesUP API
  description: |
    API REST du catalogue et des parcours clients GamesUP.
    Les ressources personnelles sont toujours déterminées par le jeton JWT :
    aucune route ne permet de sélectionner un autre utilisateur.
  contact:
    name: GamesUP
  license:
    name: Projet pédagogique
  version: 1.0.0
servers:
- url: /
  description: Serveur courant
...
paths:
...
  /api/v1/admin/catalog/games/{id}:
    get:
      tags:
      - Catalog administration
      summary: "Consulter un jeu, actif ou archivé"
      operationId: findGame
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdminGameResponse"
        "400":
          description: Requête ou paramètre invalide.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: "JWT absent, invalide ou expiré."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Rôle ou propriété de la ressource insuffisant.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Ressource introuvable.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "409":
          description: Conflit avec l'état courant.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - bearerAuth: []
    put:
      tags:
      - Catalog administration
      summary: Remplacer les données éditables d'un jeu
      operationId: updateGame
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminGameUpdateRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdminGameResponse"
        "400":
          description: Requête ou paramètre invalide.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: "JWT absent, invalide ou expiré."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Rôle ou propriété de la ressource insuffisant.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Ressource introuvable.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "409":
          description: Conflit avec l'état courant.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - bearerAuth: []
    delete:
      tags:
      - Catalog administration
      summary: Archiver un jeu
      description: L'archivage est logique et préserve l'historique.
      operationId: archiveGame
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: No Content
        "400":
          description: Requête ou paramètre invalide.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: "JWT absent, invalide ou expiré."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Rôle ou propriété de la ressource insuffisant.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Ressource introuvable.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "409":
          description: Conflit avec l'état courant.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - bearerAuth: []
...
  /api/v1/auth/register:
    post:
      tags:
      - Authentication
      summary: Créer un compte client
      description: Crée exclusivement un compte CLIENT et retourne immédiatement un
        JWT.
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
        required: true
      responses:
        "201":
          description: Compte créé.
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthenticationResponse"
              example:
                accessToken: eyJ...
                tokenType: Bearer
                expiresInSeconds: 3600
                userId: 42
                email: alice@example.com
                role: CLIENT
        "400":
          description: Requête ou paramètre invalide.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "409":
          description: Conflit avec l'état courant.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
  /api/v1/auth/login:
    post:
      tags:
      - Authentication
      summary: Se connecter
      description: Authentifie un compte actif sans révéler si l'email ou le mot de
        passe est incorrect.
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthenticationResponse"
        "400":
          description: Requête ou paramètre invalide.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "409":
          description: Conflit avec l'état courant.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
...
  /api/v1/games:
    get:
      tags:
      - Games
      summary: Rechercher les jeux
      description: |
        Retourne une page stable de jeux actifs. Les tris autorisés sont :
        id, name, price, minAge, durationMinutes et editionNumber.
      operationId: search
      parameters:
      - name: q
        in: query
        required: false
        schema:
          maxLength: 200
          minLength: 0
          type: string
      - name: category
        in: query
        required: false
        schema:
          maxLength: 150
          minLength: 0
          type: string
      - name: minPrice
        in: query
        required: false
        schema:
          minimum: 0.0
          exclusiveMinimum: false
          type: number
      - name: maxPrice
        in: query
        required: false
        schema:
          minimum: 0.0
          exclusiveMinimum: false
          type: number
      - name: page
        in: query
        required: false
        schema:
          minimum: 0
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          maximum: 100
          minimum: 1
          type: integer
          format: int32
          default: 20
      - name: sort
        in: query
        required: false
        schema:
          maxLength: 50
          minLength: 0
          type: string
          default: "name,asc"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseGameSummaryResponse"
        "400":
          description: Requête ou paramètre invalide.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Ressource introuvable.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
  /api/v1/games/{id}:
    get:
      tags:
      - Games
      summary: Consulter un jeu actif
      operationId: findById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/GameDetailResponse"
        "400":
          description: Requête ou paramètre invalide.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Ressource introuvable.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
...
  /api/v1/admin/users/{userId}:
    get:
      tags:
      - User administration
      summary: Consulter un utilisateur
      operationId: findById_1
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Requête ou paramètre invalide.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: "JWT absent, invalide ou expiré."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "403":
          description: Rôle ou propriété de la ressource insuffisant.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "404":
          description: Ressource introuvable.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
      security:
      - bearerAuth: []
...
components:
  schemas:
    GameSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        price:
          type: number
        minPlayers:
          type: integer
          format: int32
        maxPlayers:
          type: integer
          format: int32
        minAge:
          type: integer
          format: int32
        durationMinutes:
          type: integer
          format: int32
        publisher:
          type: string
        authors:
          type: array
          items:
            type: string
        categories:
          type: array
          items:
            type: string
...
    CatalogReferenceResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    AdminGameUpdateRequest:
      required:
      - authorIds
      - categoryIds
      - description
      - durationMinutes
      - editionNumber
      - maxPlayers
      - minAge
      - minPlayers
      - name
      - price
      - publisherId
      type: object
      properties:
        name:
          maxLength: 200
          minLength: 0
          type: string
        description:
          maxLength: 4000
          minLength: 0
          type: string
        price:
          minimum: 0.00
          exclusiveMinimum: false
          type: number
        minPlayers:
          minimum: 1
          type: integer
          format: int32
        maxPlayers:
          minimum: 1
          type: integer
          format: int32
        minAge:
          minimum: 0
          type: integer
          format: int32
        durationMinutes:
          minimum: 1
          type: integer
          format: int32
        editionNumber:
          minimum: 1
          type: integer
          format: int32
        publisherId:
          type: integer
          format: int64
        authorIds:
          uniqueItems: true
          type: array
          items:
            type: integer
            format: int64
        categoryIds:
          uniqueItems: true
          type: array
          items:
            type: integer
            format: int64
        playerRangeValid:
          type: boolean
    AdminGameResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        price:
          type: number
        minPlayers:
          type: integer
          format: int32
        maxPlayers:
          type: integer
          format: int32
        minAge:
          type: integer
          format: int32
        durationMinutes:
          type: integer
          format: int32
        editionNumber:
          type: integer
          format: int32
        active:
          type: boolean
        publisher:
          $ref: "#/components/schemas/CatalogReferenceResponse"
        authors:
          type: array
          items:
            $ref: "#/components/schemas/CatalogReferenceResponse"
        categories:
          type: array
          items:
            $ref: "#/components/schemas/CatalogReferenceResponse"
...
    AuthenticationResponse:
      type: object
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
        expiresInSeconds:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        email:
          type: string
        role:
          type: string
          enum:
          - CLIENT
          - ADMIN
    RegisterRequest:
      required:
      - email
      - firstName
      - lastName
      - password
      type: object
      properties:
        email:
          maxLength: 320
          minLength: 0
          type: string
          example: alice@example.com
        password:
          maxLength: 72
          minLength: 12
          type: string
          format: password
          writeOnly: true
          example: A-strong-password-2026
        firstName:
          maxLength: 100
          minLength: 0
          type: string
          example: Alice
        lastName:
          maxLength: 100
          minLength: 0
          type: string
          example: Martin
    LoginRequest:
      required:
      - email
      - password
      type: object
      properties:
        email:
          maxLength: 320
          minLength: 0
          type: string
          example: alice@example.com
        password:
          maxLength: 72
          minLength: 0
          type: string
          format: password
          writeOnly: true
          example: A-strong-password-2026
...
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum:
          - CLIENT
          - ADMIN
        enabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        properties:
          type: object
          additionalProperties:
            type: object
...
    PageResponseGameSummaryResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/GameSummaryResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
    GameDetailResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        price:
          type: number
        minPlayers:
          type: integer
          format: int32
        maxPlayers:
          type: integer
          format: int32
        minAge:
          type: integer
          format: int32
        durationMinutes:
          type: integer
          format: int32
        editionNumber:
          type: integer
          format: int32
        publisher:
          type: string
        authors:
          type: array
          items:
            type: string
        categories:
          type: array
          items:
            type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      description: Jeton retourné par l'inscription ou la connexion.
      scheme: bearer
      bearerFormat: JWT